+++ R/shoji-order.R
@@ -418,6 +418,32 @@ setdiff_entities <- function (x, ents, remove.na=FALSE) {
     return(x)
 }
 
+intersect_entities <- function (x, ents, remove.na=TRUE) {
+    ## Keep only the part of x (Order) containing "ents" (entity references)
+    if (!is.character(ents)) {
+        ## Get just the entity URLs
+        ents <- urls(ents)
+    }
+
+    if (inherits(x, "ShojiOrder") || inherits(x, "OrderGroup")) {
+        entities(x) <- intersect_entities(entities(x), ents)
+    } else if (is.list(x)) {
+        ## We're inside entities, which may have nested groups
+        grps <- vapply(x, inherits, logical(1), what="OrderGroup")
+        x[grps] <- lapply(x[grps], intersect_entities, ents)
+        matches <- unlist(x[!grps]) %in% ents
+        if (any(!matches)) {
+            ## Put in NAs so that any subsequent assignment into this object
+            ## assigns into the right position. Then strip NAs after
+            x[!grps][!matches] <- rep(list(NA_character_), sum(!matches))
+        }
+    }
+    if (remove.na) {
+        x <- removeMissingEntities(x)
+    }
+    return(x)
+}
+
 removeMissingEntities <- function (x) {
     ## Remove NA entries, left by setdiff_entities, from @graph/entities
     if (inherits(x, "ShojiOrder") || inherits(x, "OrderGroup")) {
@@ -432,6 +458,84 @@ removeMissingEntities <- function (x) {
         }
     }
     return(x)
+}
+
+#' Remove OrderGroups with no entities
+#'
+#' This function recurses through a \code{ShojiOrder}/\code{OrderGroup} and
+#' removes any groups that contain no entities.
+#'
+#' @param x VariableOrder, DatasetOrder, VariableGroup, or DatasetGroup
+#' @return \code{x} with empty groups removed.
+#' @export
+removeEmptyGroups <- function (x) {
+    if (inherits(x, "ShojiOrder") || inherits(x, "OrderGroup")) {
+        entities(x) <- removeEmptyGroups(entities(x))
+    } else if (is.list(x)) {
+        ## We're inside entities, which may have nested groups
+        grps <- vapply(x, inherits, logical(1), what="OrderGroup")
+        if (any(grps)) {
+            empties <- vapply(x[grps], function (g) length(urls(g)) == 0,
+                logical(1))
+            ## Recurse through non-empty groups
+            if (any(!empties)) {
+                nonempty <- which(grps)[!empties]
+                x[nonempty] <- lapply(x[nonempty], removeEmptyGroups)
+            }
+            ## Drop empty groups
+            if (any(empties)) {
+                x <- x[-which(grps)[empties]]
+            }
+        }
+    }
+    return(x)
+}
+
+#' Remove duplicated entites from an order/group
+#'
+#' This function recurses through a \code{ShojiOrder}/\code{OrderGroup} and
+#' removes any duplicate entities. As with the default of
+#' \code{\link[base]{duplicated}}, the first appearance of an entity is kept,
+#' and subsequent occurences are marked as duplicated and removed. (Unlike
+#' \code{duplicated}, however, there is no option to reverse that order.)
+#' "First" occurence of an entity is determined by the function's recursion:
+#' within each group, nested groups are processed first, in order, and
+#' recursively their nested groups are processed. See the test suite, in
+#' test-variable-order.R, for an example that illustrates which entities are
+#' dropped as duplicate.
+#'
+#' @param x VariableOrder, DatasetOrder, VariableGroup, or DatasetGroup
+#' @return \code{x} with duplicate entities removed.
+#' @seealso \code{\link{duplicates}}, which when set to \code{FALSE} also calls
+#' this function.
+#' @export
+dedupeOrder <- function (x) {
+    ## Collect seen urls outside, diff out urls, recurse into groups, update seen urls
+    seen <- c()
+
+    .dedupe <- function (x) {
+        if (inherits(x, "ShojiOrder") || inherits(x, "OrderGroup")) {
+            entities(x) <- .dedupe(entities(x))
+        } else if (is.list(x)) {
+            ## We're inside entities, which may have nested groups
+            grps <- vapply(x, inherits, logical(1), what="OrderGroup")
+
+            ## First, recurse through groups:
+            x[grps] <- lapply(x[grps], .dedupe)
+
+            ## Then, dedupe URLs at this level, and drop any that we've already seen
+            ents <- unlist(x[!grps])
+            badents <- duplicated(ents) | ents %in% seen
+            if (any(badents)) {
+                x <- x[-which(!grps)[badents]]
+            }
+
+            ## Update "seen" with the URLs we didn't drop
+            seen <<- c(seen, ents[!badents])
+        }
+        return(x)
+    }
+    return(.dedupe(x))
 }
 
 #' Get un(grouped) OrderGroups
@@ -470,7 +574,7 @@ ungrouped <- function (order.obj) {
 #'}
 #' @export
 moveToGroup <- function (x, value) {
-    if (!inherits(value, "ShojiGroup")) {
+    if (!inherits(value, "OrderGroup")) {
         ## If it's a Group, let's move it as is. If not, get the URLs
         ## TODO: this won't do the right thing for moving Dataset to DatasetGroup
         value <- urls(value)