+++ R/show.R
@@ -277,3 +277,7 @@ setMethod("getShowContent", "CrunchFilter",
 #' @rdname show-crunch
 #' @export
 setMethod("show", "CrunchCube", function (object) show(cubeToArray(object)))
+
+#' @rdname show-crunch
+#' @export
+setMethod("show", "OrderGroup", function (object) cat(showOrderGroup(object, index=structure(lapply(urls(object), function (x) list(name=x)), .Names=urls(object))), sep="\n"))