+++ NEWS.md
@@ -1,4 +1,5 @@
 ### crunch 1.13.5 (under development)
+* More utility functions for working with order objects: `dedupeOrder`, `removeEmptyGroups`
 * `appendDataset` can now append a subset of variables.
 
 ### crunch 1.13.4